Move contents of ExhumeHadronizer.{h,cc} to ExhumeGeneratorFilter
