+++ GeneratorInterface/ExhumeInterface/plugins/DiPhoton.cpp
@@ -3,7 +3,7 @@
 //   Written by James Monk and Andrew Pilkington
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "GeneratorInterface/ExhumeInterface/interface/DiPhoton.h"
+#include "DiPhoton.h"
 
 ////////////////////////////////////////////////////////////////////////////////
 Exhume::DiPhoton::DiPhoton(const edm::ParameterSet& pset) : TwoSpace(pset) {
