+++ GeneratorInterface/ExhumeInterface/plugins/GG.cpp
@@ -3,7 +3,7 @@
 //   Written by James Monk and Andrew Pilkington
 //////////////////////////////////////////////////////////////////////////////
 
-#include "GeneratorInterface/ExhumeInterface/interface/GG.h"
+#include "GG.h"
 
 //////////////////////////////////////////////////////////////////////////////
 Exhume::GG::GG(const edm::ParameterSet& pset) : TwoSpace(pset) {
