+++ GeneratorInterface/ExhumeInterface/plugins/QQ.cpp
@@ -3,7 +3,7 @@
 //   Written by James Monk and Andrew Pilkington
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "GeneratorInterface/ExhumeInterface/interface/QQ.h"
+#include "QQ.h"
 
 ////////////////////////////////////////////////////////////////////////////////
 Exhume::QQ::QQ(const edm::ParameterSet& pset) : TwoSpace(pset) {
