+++ GeneratorInterface/ExhumeInterface/plugins/TwoSpace.cpp
@@ -1,4 +1,4 @@
-#include "GeneratorInterface/ExhumeInterface/interface/TwoSpace.h"
+#include "TwoSpace.h"
 #include "CLHEP/Random/RandomEngine.h"
 
 Exhume::TwoSpace::TwoSpace(const edm::ParameterSet& pset) : CrossSection(pset) {
